+++ pieplot/requirements.txt
@@ -10,6 +10,7 @@ dash-table==3.6.0
 decorator==4.4.0
 Flask==1.0.2
 Flask-Compress==1.4.0
+gunicorn==19.9.0
 idna==2.8
 ipython-genutils==0.2.0
 itsdangerous==1.1.0
@@ -20,7 +21,7 @@ MarkupSafe==1.1.1
 nbformat==4.4.0
 numpy==1.16.3
 pandas==0.24.2
-plotly==3.8.1
+plotly==3.9.0
 pyrsistent==0.15.1
 python-dateutil==2.8.0
 pytz==2019.1
